services/ingestion-service/requirements.txt
fastapi
uvicorn[standard]
tiktoken
python-dotenv
openai
psycopg[binary]